Various changes from PR
