+++ api/pom.xml
@@ -47,11 +47,6 @@
             <version>${springboot.version}</version>
             <scope>test</scope>
         </dependency>
-        <dependency>
-            <groupId>com.jayway.jsonpath</groupId>
-            <artifactId>json-path</artifactId>
-            <version>2.4.0</version>
-        </dependency>
     </dependencies>
 
     <build>
