+++ api/src/main/java/gov/cms/ab2d/api/controller/BulkDataAccessAPI.java
@@ -1,9 +1,12 @@
 package gov.cms.ab2d.api.controller;
 
+import com.fasterxml.jackson.databind.JsonNode;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import gov.cms.ab2d.api.service.JobService;
 import gov.cms.ab2d.api.util.Constants;
 import gov.cms.ab2d.domain.Job;
 import io.swagger.annotations.*;
+import org.apache.commons.lang3.exception.ExceptionUtils;
 import org.hl7.fhir.dstu3.model.OperationOutcome;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.io.Resource;
@@ -11,14 +14,14 @@ import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
 import javax.validation.constraints.NotBlank;
 import java.io.IOException;
 
 import static gov.cms.ab2d.api.util.Constants.API_PREFIX;
-import static gov.cms.ab2d.api.util.FHIRUtil.getSuccessfulOutcome;
-import static gov.cms.ab2d.api.util.FHIRUtil.outcomeToJSON;
+import static gov.cms.ab2d.api.util.FHIRUtil.*;
 
 @Api(value = "Bulk Data Access API", description =
         "API through which an authenticated and authorized PDP sponsor" +
@@ -52,7 +55,7 @@ public class BulkDataAccessAPI {
     )
     @ResponseStatus(value = HttpStatus.ACCEPTED)
     @GetMapping("/Patient/$export")
-    public ResponseEntity<String> exportAllPatients(
+    public ResponseEntity<Void> exportAllPatients(
             @ApiParam(value = "String of comma-delimited FHIR resource types. Only resources of " +
                     "the specified resource types(s) SHALL be included in the response.",
                     allowableValues = "ExplanationOfBenefits")
@@ -74,10 +77,7 @@ public class BulkDataAccessAPI {
         HttpHeaders responseHeaders = new HttpHeaders();
         responseHeaders.add("Content-Location", statusURL);
 
-        OperationOutcome operationOutcome = getSuccessfulOutcome(String.format("Request %s accepted for processing", job.getJobID()));
-
-        String encoded = outcomeToJSON(operationOutcome);
-        return new ResponseEntity<>(encoded, responseHeaders,
+        return new ResponseEntity<>(null, responseHeaders,
                 HttpStatus.ACCEPTED);
     }
 
@@ -139,4 +139,17 @@ public class BulkDataAccessAPI {
             throw new RuntimeException("The current user has an active job");
         }
     }*/
+
+    @ControllerAdvice
+    class ErrorHandler extends ResponseEntityExceptionHandler {
+
+        @ExceptionHandler(Exception.class)
+        @ResponseStatus(HttpStatus.BAD_REQUEST)
+        public ResponseEntity<JsonNode> assertionException(final Exception e) throws IOException {
+            String msg = ExceptionUtils.getRootCauseMessage(e);
+            OperationOutcome operationOutcome = getErrorOutcome(msg);
+            String encoded = outcomeToJSON(operationOutcome);
+            return new ResponseEntity<>(new ObjectMapper().readTree(encoded), HttpStatus.BAD_REQUEST);
+        }
+    }
 }
